droplet-libs更新到最新
**复现步骤**

* 无

**原因和解决方案**

* 更新doc版本和sequence用uint64

**影响范围**

* 无

**验证方案**

* 验证通过

**涉及分支**

* master
* 5.5.6

**检查项**

- [ ] 需要更新依赖
- [ ] 日志打印不会造成磁盘占用空间过多
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @xiangyang

/cc @jinjie

+++ sender/marshaller.go
@@ -17,7 +17,7 @@ const (
 type ZeroDocumentMarshaller struct {
 	input    queue.QueueReader
 	outputs  []queue.QueueWriter
-	sequence uint32
+	sequence uint64
 }
 
 // NewZeroDocumentMarshaller 从input读取app.Document，转换为pb.ZeroDocument结构，复制并推送到每一个outputs里
